requirements.txt
psutil
requests
PyYAML
jsonschema
pytest
